feat: Clean + Hexagonal 아키텍쳐 생성 및 User 도메인 예시 추가
