+++ src/main/kotlin/com/chaw/shopping_note/app/user/infrastructure/repository/UserRepository.kt
@@ -1,9 +1,8 @@
-package com.chaw.shopping_note.app.infrastructure.mysql.user
+package com.chaw.shopping_note.app.user.infrastructure.repository
 
-import com.chaw.shopping_note.app.domain.user.entity.User
+import com.chaw.shopping_note.app.user.entity.User
 import org.springframework.data.repository.reactive.ReactiveCrudRepository
 import org.springframework.stereotype.Repository
-import reactor.core.publisher.Mono
 
 @Repository
 interface UserRepository : ReactiveCrudRepository<User, Long> {
